backend/requirements.txt
Flask
Flask-Cors
python-dotenv
langchain
langchain-google-genai
